Fixed almost everything.
The program now handles new lines when line width is exceeded, and then pads the words with spaces in between to make up for the remaining characters left over. Although currently, the program does not format the last line of text. There also is a bug when running on the entire works of Shakespeare, which contains a 46-letter word, and I think it gets stuck when it reaches that point. The Shakespeare bug is fixed when the line width is specified to 57 or above, rather than the standard 50 from the makefile.